+++ Contoso.Bsl/Program.cs
@@ -6,15 +6,7 @@ var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 
-builder.Services.AddControllers().AddJsonOptions
-(
-    options =>
-    {
-        options.JsonSerializerOptions.Converters.Add(new DescriptorConverter());
-        options.JsonSerializerOptions.Converters.Add(new ModelConverter());
-        options.JsonSerializerOptions.Converters.Add(new ObjectConverter());
-    }
-);
+builder.Services.AddCertificateAuthorization(builder);
 
 builder.Services
     .AddSqlServerDatabaseConfiguration(builder.Configuration.GetConnectionString("DefaultConnection")!)
@@ -27,6 +19,8 @@ var app = builder.Build();
 // Configure the HTTP request pipeline.
 
 app.UseHttpsRedirection();
+
+app.UseAuthentication();
 
 app.UseAuthorization();